requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.26.0
Django==3.2.5
django-crispy-forms==1.12.0
idna==2.10
Pillow==8.3.0
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
